Refactor: Add workflow for publishing types to NPM
This commit adds a new workflow file, "publish-types.yml", which is triggered on push events to the "production" branch and when changes are made to the "src/types/index.ts" file. The workflow checks out the repository, sets up NodeJS, installs NPM dependencies, builds the package using tsup bundler, bumps the package version using @helperkits/bumper, and finally publishes the package to NPM. The package is published under the name "@tmlmobilidade/core-types" and includes the necessary exports for importing and using the types.

New files:
- .github/workflows/publish-types.yml
- scripts/build-types.ts
- tsup-types.config.ts